services/orchestrator-api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
gitpython>=3.1.43
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.2.0
typing-extensions>=4.0.0
